Typescript-ify `sendBeacon`
